+++ sporco/cupy/admm/__init__.py
@@ -114,6 +114,12 @@ admm.cbpdn.AddMaskSim.index_primary = _index_primary
 admm.cbpdn.AddMaskSim.index_addmsk = _index_addmsk
 
 
+# Construct sporco.cupy.admm.cbpdnli
+admm.pdcsc = sporco_cupy_patch_module('sporco.admm.cbpdnli',
+                                      {'cbpdn': admm.cbpdn, 'sl': linalg,
+                                       'sp': prox})
+
+
 # Construct sporco.cupy.admm.cbpdntv
 admm.cbpdntv = sporco_cupy_patch_module('sporco.admm.cbpdntv',
                                         {'admm': admm.admm, 'cr': cnvrep,
